exercice 4.20 - 4.22

+++ Part4/bloglist/controllers/blog.js
@@ -1,17 +1,15 @@
-const jwt = require('jsonwebtoken')
 const blogRouter = require('express').Router()
 const { Blog } = require('../models/blog')
-const User = require('../models/user')
 
-const getToken = req => {
-  const auth = req.get('authorization')
-  console.log(auth)
-  if (auth && auth.startsWith('Bearer ')) {
-    console.log('getToken')
-    return auth.replace('Bearer ', '')
-  }
-  return null
-}
+// const getToken = req => {
+//   const auth = req.get('authorization')
+//   console.log(auth)
+//   if (auth && auth.startsWith('Bearer ')) {
+//     console.log('getToken')
+//     return auth.replace('Bearer ', '')
+//   }
+//   return null
+// }
 
 blogRouter.get('/', async (_request, response) => {
   const all = await Blog.find({}).populate('user')
@@ -19,19 +17,19 @@ blogRouter.get('/', async (_request, response) => {
 })
 
 blogRouter.get('/:id', async (req, res) => {
-  const note = await Blog.findById(req.params.id).populate('user')
-  if (note) {
-    res.json(note)
+  const post = await Blog.findById(req.params.id).populate('user')
+  if (post) {
+    res.json(post)
   } else {
-    res.status(404).send('message id not in db')
+    res.status(404).send('id not in db')
   }
 })
 
 blogRouter.post('/', async (request, response) => {
-  const { body } = request
-  console.log(request.get('authorization'))
-  const decodedToken = jwt.verify(getToken(request), process.env.SECRET)
-  const user = await User.findById(decodedToken.id)
+  const { body, user } = request
+  if (!user) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -46,8 +44,21 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (req, res) => {
-  const deleted = await Blog.findByIdAndDelete(req.params.id)
-  res.status(204).json(deleted)
+  const post = await Blog.findById(req.params.id)
+  const { user } = req
+  if (!post) {
+    return res.status(404).json({ error: 'post not in db' })
+  }
+  if (post.user.toString() === user.id.toString()) {
+    const deleted = await Blog.findByIdAndDelete(req.params.id)
+    user.blogPosts = user.blogPosts.filter(
+      id => id.toString() !== req.params.id
+    )
+    await user.save()
+    console.log(user)
+    return res.status(204).json(deleted)
+  }
+  return res.status(401).json({ error: 'not allowed' })
 })
 
 blogRouter.put('/:id', async (req, res) => {
